+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [BREAKING] Replaced the `api-bind-url` param for `api-bind-port` ([#156](https://github.com/0xMiden/miden-faucet/pull/156)).
 - [BREAKING] Replaced the `frontend-url` param for `frontend-bind-port` ([#156](https://github.com/0xMiden/miden-faucet/pull/156)).
 - [BREAKING] Added `no-frontend` param to optionally disable the frontend server ([#156](https://github.com/0xMiden/miden-faucet/pull/156)).
+- Redesigned the frontend ([#201](https://github.com/0xMiden/miden-faucet/pull/201)).
 
 ## 0.12.4 (2025-12-04)